backend/pyproject.toml
[project]
name = "notes-graph"
version = "0.1.0"
description = "FastAPI backend for notes as a graph"
readme = "../README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "pydantic",
    "sqlalchemy[asyncio]",
    "asyncpg",
    "python-dotenv",
]

[tool.uv]
dev-dependencies = [
    "pytest",
    "pytest-asyncio",
    "httpx",
    "ruff",
]
